Fix Copy path failing in desktop app by using native clipboard first
